+++ examples/advanced/src/gen/models/ts/tag/Tag.ts
@@ -1,5 +1,12 @@
+/**
+ * @deprecated
+ */
 export type TagTag = {
   /**
+   * @deprecated
+   * @minLength 5
+   * @maxLength 7
+   * @default 1
    * @type integer | undefined, int64
    */
   id?: number