+++ examples/advanced/src/gen/zod/tag/tagSchema.ts
@@ -3,7 +3,7 @@ import type { ToZod } from '@kubb/plugin-zod/utils'
 import { z } from 'zod'
 
 export const tagTagSchema = z.object({
-  id: z.number().int().optional(),
+  id: z.number().int().min(5).max(7).default(1),
   name: z.string().optional(),
 } satisfies ToZod<TagTag>)